requirements.txt
# NOTE: this minimalist file is only used for colext, so that we create containers with the bare minimum number of requirements
hydra-core==1.3.2
numpy
torch
wandb
